Format timestamps more tersely in the notification panel
